+++ app/src/main/java/com/example/appraisal/UI/main_menu/forum/ForumHomeActivity.java
@@ -1,20 +1,18 @@
-package com.example.appraisal.UI;
+package com.example.appraisal.UI.main_menu.forum;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
-import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.ListView;
 
-import androidx.appcompat.app.AppCompatActivity;
-
 import com.example.appraisal.R;
+import com.example.appraisal.UI.main_menu.MainMenuCommonActivity;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 
-public class ForumHomeActivity extends AppCompatActivity {
+public class ForumHomeActivity extends MainMenuCommonActivity {
 
     private ListView forum_list;
     private ArrayList<String> forum_content;
