+++ app/src/main/res/layout/activity_forum_home.xml
@@ -58,39 +58,39 @@
         app:layout_constraintStart_toStartOf="parent">
 
         <Button
-            android:id="@+id/home_f"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_weight="1"
-            android:text="home"/>
+                android:id="@+id/home_button"
+                android:layout_width="wrap_content"
+                android:layout_height="match_parent"
+                android:layout_weight="1"
+                android:text="home" android:onClick="toHome"/>
 
         <Button
-            android:id="@+id/search_f"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_weight="1"
-            android:text="search" />
+                android:id="@+id/search_button"
+                android:layout_width="wrap_content"
+                android:layout_height="match_parent"
+                android:layout_weight="1"
+                android:text="search" android:onClick="toSearch"/>
 
         <Button
-            android:id="@+id/forum_f"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_weight="1"
-            android:text="forum" />
+                android:id="@+id/forum_button"
+                android:layout_width="wrap_content"
+                android:layout_height="match_parent"
+                android:layout_weight="1"
+                android:text="forum" android:onClick="toForum"/>
 
         <Button
-            android:id="@+id/myExperiment_f"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_weight="1"
-            android:text="list" />
+                android:id="@+id/my_exp_button"
+                android:layout_width="wrap_content"
+                android:layout_height="match_parent"
+                android:layout_weight="1"
+                android:text="list" android:onClick="toMyExps"/>
 
         <Button
-            android:id="@+id/profile_f"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_weight="1"
-            android:text="profile" />
+                android:id="@+id/profile_button"
+                android:layout_width="wrap_content"
+                android:layout_height="match_parent"
+                android:layout_weight="1"
+                android:text="profile" android:onClick="toProfile"/>
 
     </LinearLayout>
