+++ app/src/main/res/layout/activity_forum_replies.xml
@@ -48,39 +48,39 @@
         app:layout_constraintStart_toStartOf="parent">
 
         <Button
-            android:id="@+id/home_f_r"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_weight="1"
-            android:text="home"/>
+                android:id="@+id/home_button"
+                android:layout_width="wrap_content"
+                android:layout_height="match_parent"
+                android:layout_weight="1"
+                android:text="home" android:onClick="toHome"/>
 
         <Button
-            android:id="@+id/search_f_r"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_weight="1"
-            android:text="search" />
+                android:id="@+id/search_button"
+                android:layout_width="wrap_content"
+                android:layout_height="match_parent"
+                android:layout_weight="1"
+                android:text="search" android:onClick="toSearch"/>
 
         <Button
-            android:id="@+id/forum_f_r"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_weight="1"
-            android:text="forum" />
+                android:id="@+id/forum_button"
+                android:layout_width="wrap_content"
+                android:layout_height="match_parent"
+                android:layout_weight="1"
+                android:text="forum" android:onClick="toForum"/>
 
         <Button
-            android:id="@+id/myExperiment_f_r"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_weight="1"
-            android:text="list" />
+                android:id="@+id/my_exp_button"
+                android:layout_width="wrap_content"
+                android:layout_height="match_parent"
+                android:layout_weight="1"
+                android:text="list" android:onClick="toMyExps"/>
 
         <Button
-            android:id="@+id/profile_f_r"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_weight="1"
-            android:text="profile" />
+                android:id="@+id/profile_button"
+                android:layout_width="wrap_content"
+                android:layout_height="match_parent"
+                android:layout_weight="1"
+                android:text="profile" android:onClick="toProfile"/>
 
     </LinearLayout>
