+++ app/src/main/res/layout/activity_selection.xml
@@ -4,7 +4,7 @@
                                                    xmlns:tools="http://schemas.android.com/tools"
                                                    android:layout_width="match_parent"
                                                    android:layout_height="match_parent"
-    tools:context=".UI.SelectionActivity">
+    tools:context=".UI.main_menu.SelectionActivity">
     >
 
     <Button
