+++ app/src/main/res/layout/activity_subscription.xml
@@ -25,21 +25,21 @@
         app:layout_constraintBottom_toBottomOf="parent">
 
         <Button
-            android:id="@+id/home"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_weight="1"
-            android:text="home" />
+                android:id="@+id/home_button"
+                android:layout_width="wrap_content"
+                android:layout_height="match_parent"
+                android:layout_weight="1"
+                android:text="home" android:onClick="toHome"/>
 
         <Button
-            android:id="@+id/search"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_weight="1"
-            android:text="search" />
+                android:id="@+id/search_button"
+                android:layout_width="wrap_content"
+                android:layout_height="match_parent"
+                android:layout_weight="1"
+                android:text="search" android:onClick="toSearch"/>
 
         <Button
-            android:id="@+id/forum"
+            android:id="@+id/forum_button"
             android:layout_width="wrap_content"
             android:layout_height="match_parent"
             android:layout_weight="1"
@@ -47,7 +47,7 @@
             android:text="forum" />
 
         <Button
-            android:id="@+id/myExperiment"
+            android:id="@+id/my_exp_button"
             android:layout_width="wrap_content"
             android:layout_height="match_parent"
             android:layout_weight="1"
@@ -55,11 +55,11 @@
             android:text="list" />
 
         <Button
-            android:id="@+id/profile"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_weight="1"
-            android:text="profile" />
+                android:id="@+id/profile_button"
+                android:layout_width="wrap_content"
+                android:layout_height="match_parent"
+                android:layout_weight="1"
+                android:text="profile" android:onClick="toProfile"/>
     </LinearLayout>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
